requirements.txt
aleph_alpha_client==7.0.0
gradio>=4.14.0
lingua-language-detector==2.0.2
nlpaug==1.1.11
nltk==3.9
numpy>=1.24.0
pydantic==2.6.1
pytest>=7.4.3
python-dotenv==1.0.1
requests>=2.32.2
scikit-learn>=1.3.0
scipy==1.12.0
spacy==3.8.5
textdistance==4.6.1
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
pandas>=2.0.0
matplotlib==3.8.2
datasets>=2.14.0
tqdm>=4.66.1
huggingface-hub>=0.5.0
torch>=2.0.0
transformers>=4.36.0
pyarrow>=14.0.1
psutil>=5.9.0
accelerate>=0.25.0
langdetect==1.0.9
# en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
# de_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl
bitsandbytes==0.41.1
# For UI
tk
